Added the Truck class to the vehicle package.
Removed the timeSpent field and the act(), spendTime() methods from the
Vehicle class.

Changed the access-modifier of the fill() method in the Vehicle class
from private to public.

Override the equals(Object) and the toString() methods in the Vehicle,
SmallCar, Motorbike, FamilySedan and Truck classes.

Added documentation to SmallCar, Motorbike, FamilySedan and Truck
classes.

Removed the getJFrame method from the SimulationView interface, added a
Show(Station) method to the SimulationView interface.

Added the constructor to the CommandLine class.