Created the Animated class for showing  visual representations of the simulation.
Added a protected method int getFuelLevel() in the Vehicle class to
allow subclasses to access the fuel level.

Added an abstract Vehicle clone() method into the Vehicle class to
create deep clones of the Vehicle subclasses.

Added a protected constructor to the Vehicle class to allow subclasses
to call the Vehicle construction to create identical clones of the
vehicle subclass.

Added a private constructor into the SmallCar, Motorbike, FamilySedan
and Truck classes to be used when cloning the Vehicle subclasses.

Override the Vehicle superclass clone method in the SmallCar, Motorbike,
FamilySedan and Truck classes to create deep clones.

Implemented test cases for the SmallCar, Motorbike, FamilySedan and
Truck classes.