Development and Maintenance (28/03/17)
[WARNINGS]
1. Documentation for Station.java is incomplete

[CHANGES]
1. Removed canContain(RoadUser) from Station.java as changes to
enter(RoadUser) in Station.java negate its existence.
2. Changed type of fuel and sales statistic fields in Station.java from
int/double to Map<Class<? extends RoadUser>, Double> to allow the
station to record these statistics in relation to the type of their
RoadUser.
3. Added sum(Map<Class<? extends RoadUser>, Double>) in Station.java
which sums the contents of the statistic fields Maps to give statistics
like fuelProfit that are global to the station.
4. Added update(Map<Class<? extends RoadUser>, Double>, Class<? extends
RoadUser>, Double) to Station.java which adds a specified Double to the
statistic fields of a road user type.
5. Modified clone(), toString() and equals() to reflect these
aforementioned changes.